pyproject.toml
[tool.poetry]
name = "datapipe-example-realtime-api"
version = "0.1.0"
description = ""
authors = ["Andrey Tatarinov <a@tatarinov.co>"]

[tool.poetry.dependencies]
python = "^3.8"
datapipe-core = "0.11.0-beta.5"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}

[tool.poetry.dev-dependencies]
mypy = "^0.961"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
